Eventos y sectores guardandose en bd correctamente
